requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
fastapi==0.73.0
frozenlist==1.3.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
multidict==6.0.2
pydantic==1.9.0
python-dotenv==0.19.2
rfc3986==1.5.0
slack_bolt==1.11.4
slack-sdk==3.14.1
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.4
yarl==1.7.2